Undo "utilities" command, implement auto listener
